Create Data
The dataset used in this study includes posts from the r/Depression and r/Anxiety subreddits between 2019 and 2022. The dataset comprises of 351,409 unique users and their posts. Due to the sensitive nature of the data, the original dataset is not publicly available. However, a subset of preprocessed data used in this study is included in the repository.